Обновил MVP онбординга
